requirements.txt
PyOpenGL
PyOpenGL_accelerate
PySide6; python_version > '3.9'
PySide2; python_version <= '3.9'
accelerate
huggingface-hub
tokenizers
transformers
colorama
PyGLM
numba
glfw
assimp_py
antlr4-python3-runtime==4.9.3
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.0
cmake==3.27.7
decorator==5.1.1
diffusers==0.26.3
einops==0.7.0
executing==2.0.0
filelock==3.12.4
fsspec==2023.9.2
idna==3.4
importlib-metadata==6.8.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
lit==17.0.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.8.1.78
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyre-extensions==0.0.23
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
torch==2.0.0
torchvision==0.15.1
tqdm==4.66.1
traitlets==5.11.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.8
xformers==0.0.17
zipp==3.17.0